requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
ultralytics>=8.0.0
onnx>=1.14.0
onnxruntime-gpu>=1.16.0; sys_platform != 'darwin'
onnxruntime>=1.16.0; sys_platform == 'darwin'
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
PyYAML>=6.0
filterpy>=1.4.5
scipy>=1.10.0
